refactor/cleanup migration iterative network send
